// client/src/store/actions/ActionTypes.d.ts
export enum AuthActions {
    STORE_TOKEN = 'STORE_TOKEN',
    AUTH_SUCCESS = 'AUTH_SUCCESS',
    AUTH_FAILED = 'AUTH_FAILED',
    AUTH_INIT = 'AUTH_INIT',
}

export enum TourActions {
    FETCH_USER_TOURS_INIT = 'FETCH_USER_TOURS_INIT',
    FETCH_USER_TOURS_SUCCESS = 'FETCH_USER_TOURS_SUCCESS',
    FETCH_USER_TOURS_ERROR = 'FETCH_USER_TOURS_ERROR',

    FETCH_TOUR_INIT = 'FETCH_TOUR_INIT',
    FETCH_TOUR_SUCCESS = 'FETCH_TOUR_SUCCESS',
    FETCH_TOUR_ERROR = 'FETCH_TOUR_ERROR',

    ADD_TOUR_INIT = 'ADD_TOUR_INIT',
    ADD_TOUR_SUCCESS = 'ADD_TOUR_SUCCESS',
    ADD_TOUR_ERROR = 'ADD_TOUR_ERROR',
}